requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.1
django-cors-headers==3.0.2
django-filter==2.1.0
django-tinymce==2.8.0
djangorestframework==3.9.4
drf-extensions==0.5.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.1.0
PyMySQL==0.9.3
pytz==2019.2
requests==2.22.0
uritemplate==3.0.0
urllib3==1.25.3
